requirements.txt
pathos
pandas
openai
python-decouple
scipy
pyyaml
matplotlib
torch
transformers
flask
xlsxwriter